[Add] server-client API

+++ tasks-server/src/tasks/tasks.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Task } from './task.entity';
@@ -6,29 +6,43 @@ import { Task } from './task.entity';
 @Injectable()
 export class TasksService {
   constructor(
-    @InjectRepository(Task) // Використовуйте Task тут
-    private tasksRepository: Repository<Task>, // Змінюємо на tasksRepository
+    @InjectRepository(Task)
+    private tasksRepository: Repository<Task>,
   ) {}
 
-  findAll(): Promise<Task[]> {
+  async findAll(): Promise<Task[]> {
     return this.tasksRepository.find();
   }
 
-  findOne(id: number): Promise<Task> {
-    return this.tasksRepository.findOneBy({ id });
+  async findOne(id: number): Promise<Task> {
+    const task = await this.tasksRepository.findOneBy({ id });
+    if (!task) {
+      throw new NotFoundException('Task not found');
+    }
+    return task;
   }
 
-  create(task: Partial<Task>): Promise<Task> {
+  async create(task: Partial<Task>): Promise<Task> {
     const newTask = this.tasksRepository.create(task);
     return this.tasksRepository.save(newTask);
   }
 
   async update(id: number, task: Partial<Task>): Promise<Task> {
+    const existingTask = await this.findOne(id);
+    if (!existingTask) {
+      throw new NotFoundException('Task not found');
+    }
+
     await this.tasksRepository.update(id, task);
     return this.findOne(id);
   }
 
   async remove(id: number): Promise<void> {
+    const task = await this.findOne(id);
+    if (!task) {
+      throw new NotFoundException('Task not found');
+    }
+
     await this.tasksRepository.delete(id);
   }
 }
